XP-694 Move XmlContentTypeParser to core-schema module
-Moved XmlContentTypeParser to core-schema module. Moved it under com.enonic.xp.core.impl.schema.content package. Also moved the associated tests.
