+++ modules/core-schema/src/main/java/com/enonic/xp/core/impl/schema/content/ContentTypeLoader.java
@@ -20,7 +20,6 @@ import com.enonic.xp.module.ModuleKey;
 import com.enonic.xp.schema.content.ContentType;
 import com.enonic.xp.schema.content.ContentTypeName;
 import com.enonic.xp.schema.content.ContentTypes;
-import com.enonic.xp.xml.parser.XmlContentTypeParser;
 
 final class ContentTypeLoader
 {
