+++ modules/core-schema/src/test/resources/com/enonic/xp/core/impl/schema/content/XmlContentTypeParserTest.xml
